backend/app/requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

boto3==1.41.0
botocore==1.41.0
s3transfer==0.14.0
jmespath==1.0.1

python-dotenv==1.2.1

requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11

psycopg2-binary==2.9.11

pandas==2.3.3
numpy==2.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

altair==6.0.0
narwhals==2.12.0

tabulate==0.9.0

PyJWT>=2.8.0
cryptography>=42.0.0
